# [XLA] vectorize row reduction for even row size (new version of PR #37260)