… tensorflow/compiler/xla/service/gpu/kernel_mapping_scheme.h
… tensorflow/compiler/xla/service/gpu/tests/BUILD
… tensorflow/compiler/xla/service/gpu/tests/reduction_vectorization_test.cc