This is a new version of #37260 that is rabased and include the fix to the DCHECK and convert it to a CHECK:
https://github.com/tensorflow/tensorflow/pull/37260#pullrequestreview-385503035
It also include the other small comments.
@cheshire 
@akuegel 

**Original description:**

This depend on LLVM commit: https://reviews.llvm.org/D74444 that is recently merged in XLA.

This PR vectorize row reduction for even row size.
For odd row size, every other row is unaligned. So would need a more complicated fix.
On a reduction of size 32x131070, I have no speed up for a max reduction, but I have 1.21x speed on the sum+pw reduction from the softmax in float16.

This was benchmarked on the row major "Softmax" TF operations.
The optimized graph contain 3 main operations: reduce_8 (the max reduction), fusion_1 (the exponential and sum reduction), fusion (the post normalization pw operations, ignored).
This was benchmarked with a matrix of input shape (32x131070) with float16 dtype:

| Kernel     | Before PR | After PR | Speed up |
| ---------- | ---------- | ---------- | ---------- |
| fusion_1  | 39.827us  | 33.843us | 1.18x |
| reduce_8 | 22.112us  | 22.047us | 1x |

Here is the not optimized HLO:
```
HloModule cluster_0
%max_ {
  %x = f16[] parameter(0)
  %y = f16[] parameter(1)
  ROOT %maximum = f16[] maximum(%x, %y)
}

%add_ {
  %x = f32[] parameter(0)
  %y = f32[] parameter(1)
  ROOT %add = f32[] add(%x, %y)
}

ENTRY %main {
  %arg0.1 = f16[32,131070]{1,0} parameter(0)
  %constant.3 = f16[] constant(-inf)
  %pad1 = f16[32,131070] pad(%arg0.1, %constant.3), padding=0_0_0x0_0_0
  %reduce.8 = f16[32]{0} reduce(%pad1, %constant.3), dimensions={1}, to_apply=%max_
  %broadcast.9 = f16[32,131070]{1,0} broadcast(%reduce.8), dimensions={0}
  %subtract.131070 = f16[32,131070]{1,0} subtract(%arg0.1, %broadcast.9)
  %exponential.11 = f16[32,131070]{1,0} exponential(%subtract.131070)
  %convert.12 = f32[32,131070]{1,0} convert(%exponential.11)
  %constant.13 = f32[] constant(0)
  %pad2 = f32[32,131070] pad(%convert.12, %constant.13), padding=0_0_0x0_0_0
  %reduce.18 = f32[32]{0} reduce(%pad2, %constant.13), dimensions={1}, to_apply=%add_
  %convert.19 = f16[32]{0} convert(%reduce.18)
  %broadcast.20 = f16[32,131070]{1,0} broadcast(%convert.19), dimensions={0}
  ROOT %divide.21 = f16[32,131070]{1,0} divide(%exponential.11, f16[32,131070]{1,0} %broadcast.20)
}
```