@nouiz 2020-04-02
I have many problems to just run the tests you gave. I have errors likes:

/usr/lib/gcc/x86_64-linux-gnu/7/../../../x86_64-linux-gnu/Scrt1.o: In function `_start':
(.text+0x20): undefined reference to `main'
collect2: error: ld returned 1 exit status

or like:

2020-04-02 15:47:09.542705: F tensorflow/compiler/xla/tests/client_library_test_base.cc:48] Non-OK-status: result.status() status: Not found: no platforms found could not create local client for testing
external/bazel_tools/tools/test/test-setup.sh: line 310: 44350 Aborted                 (core dumped) "${TEST_PATH}" "$@" 2>&1

or like: 

ERROR: /home/fbastien/github/tensorflow/tensorflow/compiler/xla/client/BUILD:182:1: Linking of rule '//tensorflow/compiler/xla/client:sharding_builder' failed (Exit 1)
gcc: error: missing argument to '-B'


I have this with all tests. So it looks like something is broken upstream (could be specific to my environment). I have errors on the upstream commit before mine too.

Can you gave me the command you use to reproduce the error you have and on which commit you tested it? It is a rebased version of this PR or this PR directly?

I'll continue to try to reproduce your error, but I already spend a few hours.