@nouiz 2020-04-02
2 days ago, a build problem was merged upstream that cause the error I saw. I made an issues: https://github.com/tensorflow/tensorflow/issues/38171

Now, hopefully, I can start to look at the problem you mentioned.