@nouiz 2020-04-02
After one day I'm still not able to reproduce the error. Are you able to reproduce it? Which bazel command line allow you to reproduce it?

Also, can you give me the full error message? It would be useful to know which DCHECK failed. I do not have that information in this PR or the previous one.