@akuegel 2020-04-03
> After one day I'm still not able to reproduce the error. Are you able to reproduce it? Which bazel command line allow you to reproduce it?

I ran this on f7b584639f07dcc95a405c857ce4071c93b5c9ac (my fixed version of your PR). However I only ran it internally, without specifying -c opt as I normally do so that DCHECK is actually evaluated. I am not so familiar with open source to know how you can make sure that DCHECKs are triggered. The easiest option for you might be to temporarily change the DCHECK in shape_util.cc to a CHECK (see stack trace below to see which line this is).

> Also, can you give me the full error message? It would be useful to know which DCHECK failed. I do not have that information in this PR or the previous one.

Here is the full stack trace:

F0403 00:03:26.839476    1674 shape_util.cc:426] Check failed: shape.IsArray() (s32[3], s32[3])
*** Check failure stack trace: ***
    @     0x7f9965176fcc  absl::logging_internal::LogMessage::DieIfFatal()
    @     0x7f9965176893  absl::logging_internal::LogMessage::SendToLog()
    @     0x7f9965175b3e  absl::logging_internal::LogMessage::Flush()
    @     0x7f9965177e68  absl::logging_internal::LogMessageFatal::~LogMessageFatal()
    @     0x7f9977a50770  xla::ShapeUtil::ElementsIn()
    @     0x7f99889179bf  xla::gpu::(anonymous namespace)::ComputeMaxUnrollFactor()
    @     0x7f998891ffd2  xla::gpu::IrEmitterUnnested::EmitTargetElementLoop()
    @     0x7f99888e2798  xla::gpu::IrEmitter::HandleReduce()
    @     0x7f998891a487  xla::gpu::IrEmitterUnnested::HandleReduce()
    @     0x7f9977f81546  xla::HloInstruction::Visit<>()
    @     0x7f9977f67e50  xla::PostOrderDFS<>()
    @     0x7f9977f80ccc  xla::HloInstruction::Accept<>()
    @     0x7f9989ed69f1  xla::HloComputation::Accept<>()
    @     0x7f99890d7aa1  xla::gpu::GpuCompiler::RunBackend()
    @     0x7f99819ed2ef  xla::LLVMCompiler::Compile()
    @     0x7f998971611e  xla::Service::BuildExecutables()
    @     0x7f9989718f56  xla::Service::ExecuteGraphParallel()
    @     0x7f9989cba4d6  xla::Client::ExecuteParallel()
    @     0x7f9989cb86ff  xla::Client::Execute()
    @     0x7f9989cb80a9  xla::Client::ExecuteAndTransfer()
    @     0x7f9989e7cd74  xla::ClientLibraryTestBase::ExecuteAndTransfer()
    @     0x7f9989e7da3d  xla::ClientLibraryTestBase::ComputeAndCompareLiteralWithStatus()
    @     0x7f9989e7d2a9  xla::ClientLibraryTestBase::ComputeAndCompareLiteral()
    @     0x7f9989ee16fb  xla::ClientLibraryTestBase::ComputeAndCompareR1<>()
    @     0x7f9989ed58fa  xla::(anonymous namespace)::ArithmeticTest_ArgMinR2Axis0_Test::TestBody()
    @     0x7f99661097b4  testing::internal::HandleSehExceptionsInMethodIfSupported<>()
    @     0x7f99660f62f2  testing::internal::HandleExceptionsInMethodIfSupported<>()
    @     0x7f99660e6353  testing::Test::Run()
    @     0x7f99660e69f7  testing::TestInfo::Run()
    @     0x7f99660e704c  testing::TestSuite::Run()
    @     0x7f99660ef335  testing::internal::UnitTestImpl::RunAllTests()
    @     0x7f996610e7b4  testing::internal::HandleSehExceptionsInMethodIfSupported<>()