@akuegel 2020-04-03
> After one day I'm still not able to reproduce the error. Are you able to reproduce it? Which bazel command line allow you to reproduce it?

Sorry that you had to spend so much time on this :-(
I didn't expect that this behaves so much differently in open source than internally. I should have provided the full stack trace immediately.
So here is how you can reproduce it in open source:
Change line 426 in tensorflow/compiler/xla/shape_util.cc from DCHECK to CHECK (the line that checks that shape is an array).
Then you should see the CHECK fail.
There might be a way to enable DCHECKs in open source, but I don't know how :-(