@nouiz 2020-04-03
I'm not able to have this XLA test works with upstream TF. So I can't try to reproduce this error. 
I created an issues that use Google Docker images: https://github.com/tensorflow/tensorflow/issues/38205

I just pushed a few small changes too. Local tests still compiling...