@nouiz 2020-04-08
Ok. Any idea on which GPU this was run?